+++ framework/wazuh/rule.py
@@ -379,13 +379,16 @@ def upload_rule_file(filename: str = None, content: str = None, overwrite: bool 
     try:
         if len(content) == 0:
             raise WazuhError(1112)
-
-        validate_wazuh_xml(content)
-
+        if exists(join(common.RULES_PATH, filename)):
+            full_path = join(common.RULES_PATH, filename)
+            raise WazuhError(1914)
         # If file already exists and overwrite is False, raise exception
         if not overwrite and exists(full_path):
             raise WazuhError(1905)
-        elif overwrite and exists(full_path):
+
+        validate_wazuh_xml(content)
+
+        if overwrite and exists(full_path):
             backup_file = f'{full_path}.backup'
             delete_file_with_backup(backup_file, full_path, delete_rule_file)
